Moved config elements to own file
